2.1.0: Merged bug fixes from 2.0.2.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@1798 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: 750bbe4c874f0324b8c66c2123961b7fc17e2e83

+++ core/producer/frame_producer.h
@@ -69,7 +69,7 @@ public:
 	virtual safe_ptr<frame_producer> get_following_producer() const {return frame_producer::empty();}  // nothrow
 	virtual void set_leading_producer(const safe_ptr<frame_producer>&) {}  // nothrow
 		
-	virtual int64_t nb_frames() const {return std::numeric_limits<int>::max();}
+	virtual uint32_t nb_frames() const {return std::numeric_limits<uint32_t>::max();}
 	
 	virtual safe_ptr<basic_frame> receive(int hints) = 0;
 	virtual safe_ptr<core::basic_frame> last_frame() const = 0;
